PSO - moved resource-related components into a separate module
